utags: Free the device name buffer when rmmod utags.ko
If the device name can be expanded, it will allocate a buffer to store expand
name, so need free it when removing.
If not expand, it will keep use buffer getting from "of_property_read_string"
and this buffer can not free on our driver, so allocate a new buffer for it.

Change-Id: Id1c18ac7caae426319813f48d1a67366216fe630
Reviewed-on: https://gerrit.mot.com/2260975
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Konstantin Makariev <kmakariev@motorola.com>
Submit-Approved: Jira Key