arch: arm64: Remove incomplete erratum 1542418 fix
The original fix was never merged into mainline:
https://lkml.iu.edu/hypermail/linux/kernel/1911.1/08600.html
>> Is there something I'm missing that makes this remotely plausible?
> No :-)
MediaTek then went on and picked incomplete code changes for this into
their kernel source for no reason. This should be safe for affected CPUs
as the bug apparently never was plausible, but even if it was a bug,
MTK's workarounds never were enabled for MT6853/kyoto anyway. It's not a
problem to remove it.

This reverts commit 5d6ad591c3a4f69b11780f74c38caa882707fb97.